Execute nethogs in the background
